Fix: OAuth 플로우 정리 - 재발급 호출을 auth 페이지로 이동하고 홈은 token 저장만 수행
